Refactoring of the code with little improvements in performance
